1. Implement greedy exploration: Find nearest unexplored cell outside some radius from the robot's positiom.
2. Implement bump sensor and obstacle escape.

3. The 2D projection of grid map now no longer use a single slice or row
taken from the depth map, but rather use a custom new row where each
column contains the minimum depth from the corresponding column in the
depth map. This may simplifies the map (and thus the planning and
exploring problem) when the robot is in an environment with many complex
objects (e.g. table with four legs).